Warn on bad data
The model is not suited for repeated data entries when the 'precision'
parameter is set to 0. Such data constitutes a probability 0 event in
every model of the model class, which is apparent in seemingly infinite
compression gain.

The warning statements have to be executed outside the scope of `with`
in order for the calling context to be correct. Otherwise, the warnings
originate from a call to `eval`.